refactor(ui): split ImprovedHelpModal sections into topic files
Drop ImprovedHelpModal.tsx from 1771 to 333 lines, well under the
800-line red-flag budget. The ~20 help content sections split across
three sibling files grouped by topic.

- ImprovedHelpModalSections.tsx (460): core sections — About,
  GettingStarted, Link, Cable, WiFi, Network, Gateway, DNS,
  Performance, Discovery
- ImprovedHelpModalAdvancedSections.tsx (408): Health Checks,
  Security, Troubleshooting
- ImprovedHelpModalSpecialtySections.tsx (600): inline helper
  components (TroubleshootingCategory, FeatureCard, StepCard,
  HelpContentSection, HelpTermList) plus Profiles, WiFi Survey,
  RTSP Checks, DICOM Checks, How-To, Glossary

Pure relocation - the lowercase <aboutSection /> JSX usages in the
modal shell are preserved as-is. No behavior change.
